Restructure App so that login functionality is handled differently
This patch adds the login/out handlers to the app level
component and passes the props down to child components. This
allows for the consistent rendering and rerending of child
components depending on the logedIn state.
