+++ src/components/Navigation.js
@@ -2,8 +2,30 @@ import React from 'react';
 import { Link } from 'react-router';
 import { Navbar, Nav, NavItem } from 'react-bootstrap';
 
+
 export default class Navigation extends React.Component {
+  constructor(props) {
+    super(props);
+    // might need to move oauthToken to parent scope
+    this.state = {
+      auth2ApiLoaded: null,
+      pickerApiLoaded: false,
+      logedIn: this.props.loggedIn,
+      handleSignInClick: this.props.handleSignInClick,
+      handleSignOutClick: this.props.handleSignOutClick
+    };
+  }
+
   render() {
+    console.log('rendering with props: ', this.props)
+    let savedAuth = JSON.parse(localStorage.getItem('GoogleAuth'));
+    let logInOrOut = null;
+    if (savedAuth === null) {
+      logInOrOut = <button id="signin-button" onClick={this.props.handleSignInClick}>Sign In</button>
+    } else {
+      logInOrOut = <button id="signout-button" onClick={this.props.handleSignOutClick}>Sign Out</button>
+    }
+
     return (
       <Navbar>
         <Nav>
@@ -19,6 +41,7 @@ export default class Navigation extends React.Component {
           <NavItem eventKey={4}>
             <Link to="upload">Upload</Link>
           </NavItem>
+            {logInOrOut}
         </Nav>
       </Navbar>
     );
