+++ src/components/Router.js
@@ -4,7 +4,6 @@ import { Router, Route, browserHistory, Redirect, IndexRoute } from 'react-route
 
 import App from './App';
 import Home from './Home';
-import Login from './Login';
 import Profile from './Profile';
 import Gallery from './Gallery';
 import Upload from './Upload';
@@ -17,7 +16,6 @@ import Upload from './Upload';
 const Root = () => (
   <Router history={browserHistory}>
     <Route path="/" component={App}>
-      <IndexRoute component={Login}/>
       <Route path="/home" component={Home} />
       <Route path="/profile" component={Profile} />
       <Route path="/upload" component={Upload} />
